Add metals

+++ vec3.v
@@ -4,27 +4,27 @@ import math
 
 struct Vec {
    mut:
-       i f32
-       j f32
-       k f32
+       i f64
+       j f64
+       k f64
 }
 
 // hypotenuse Calculate the displacement due to Vec
-pub fn (v Vec) hypotenuse() f32 {
+pub fn (v Vec) hypotenuse() f64 {
     return v.i * v.i + v.j * v.j + v.k * v.k
 }
 
-pub fn (v Vec) x() f32 {
+pub fn (v Vec) x() f64 {
     return v.i
 }
-pub fn (v Vec) y() f32 {
+pub fn (v Vec) y() f64 {
     return v.j
 }
-pub fn (v Vec) z() f32 {
+pub fn (v Vec) z() f64 {
     return v.k
 }       
 // length Calculate the abs/mangatude of Vec3 as a Vec
-pub fn (v Vec) length() f32 {
+pub fn (v Vec) length() f64 {
   return math.sqrt(v.hypotenuse())
 }
 
@@ -57,7 +57,7 @@ pub fn (a Vec) * (b Vec) Vec {
 }
 
 // mul_scalar Multiple a vec3 Vec by a scalar
-pub fn (a Vec) mul_scalar(t f32) Vec {
+pub fn (a Vec) mul_scalar(t f64) Vec {
   return Vec {
     a.i * t,
     a.j * t,
@@ -66,7 +66,7 @@ pub fn (a Vec) mul_scalar(t f32) Vec {
 }
 
 // div_scalar Divide Vec by a scalar
-pub fn (a Vec) div_scalar(t f32) Vec {
+pub fn (a Vec) div_scalar(t f64) Vec {
   return Vec {
     a.i / t,
     a.j / t,
@@ -74,7 +74,7 @@ pub fn (a Vec) div_scalar(t f32) Vec {
   }
 }
 // dot Calculate dot product between two vec3 Vecs
-pub fn (v Vec) dot(u Vec) f32 {
+pub fn (v Vec) dot(u Vec) f64 {
   return v.i * u.i + v.j * u.j + v.k * u.k
 }
 
@@ -111,11 +111,11 @@ pub fn (c RGB) str() string {
     return '$c.r $c.g $c.b'
 }
 
-pub fn (c RGB) mul_scalar(k f32) RGB {
+pub fn (c RGB) mul_scalar(k f64) RGB {
     return RGB{
-               u8(f32(c.r) * k),
-               u8(f32(c.g) * k),
-               u8(f32(c.b) * k)
+               u8(f64(c.r) * k),
+               u8(f64(c.g) * k),
+               u8(f64(c.b) * k)
                }
 }
        
@@ -126,7 +126,7 @@ struct Ray {
 }
 
 // at Calculate position along Ray at t.
-pub fn (r Ray) at(t f32) Vec {
+pub fn (r Ray) at(t f64) Vec {
   return r.a + r.b.mul_scalar(t)
 }
